+++ src/Pollus.Graphics/Window/BrowserWindow.cs
@@ -21,7 +21,7 @@ public partial class BrowserWindow : IWindow
 
     public bool IsOpen { get; private set; }
     public WindowOptions Options { get; private set; }
-    public Vec2<int> Size { get; set; }
+    public Vec2<uint> Size { get; set; }
     public INativeWindow? Native => null;
 
     nint nativeWindow;
@@ -30,12 +30,12 @@ public partial class BrowserWindow : IWindow
     {
         instance = this;
         Options = options;
-        Size = new Vec2<int>(options.Width, options.Height);
+        Size = new Vec2<uint>(options.Width, options.Height);
 
         EmscriptenSDL.Init(SDLInitFlags.InitVideo | SDLInitFlags.InitJoystick);
         nativeWindow = EmscriptenSDL.CreateWindow(options.Title,
             Silk.NET.SDL.Sdl.WindowposUndefined, Silk.NET.SDL.Sdl.WindowposUndefined,
-            options.Width, options.Height,
+            (int)options.Width, (int)options.Height,
             Silk.NET.SDL.WindowFlags.Resizable);
 
         IsOpen = true;
